requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Automat==20.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
build==1.2.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
coloredlogs==15.0.1
constantly==15.1.0
contourpy==1.2.0
cryptography==38.0.1
cssselect==1.1.0
ctransformers==0.2.27
cycler==0.12.1
Cython==3.0.10
dataclasses-json==0.6.4
dataclasses-json-speakeasy==0.5.11
Deprecated==1.2.14
distro==1.9.0
effdet==0.4.1
emoji==2.10.1
exceptiongroup==1.2.0
fastapi==0.110.0
filelock==3.8.0
filetype==1.2.0
flatbuffers==24.3.25
fonttools==4.50.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.29.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface==0.0.1
huggingface-hub==0.21.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
incremental==22.10.0
iniconfig==2.0.0
iopath==0.1.10
itemadapter==0.7.0
itemloaders==1.0.6
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.36
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langchainhub==0.1.15
langdetect==1.0.9
langsmith==0.1.38
layoutparser==0.3.4
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.2
matplotlib==3.8.3
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.16.0
onnxruntime==1.15.1
openai==1.14.3
opencv-python==4.9.0.80
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parsel==1.6.0
pdf2image==1.17.0
pdfminer==20191125
pdfminer.six==20231228
pdfplumber==0.11.0
pikepdf==8.14.0
pillow==10.2.0
pillow_heif==0.15.0
pluggy==1.4.0
poppler-utils==0.1.0
portalocker==2.8.2
posthog==3.5.0
Protego==0.2.1
protobuf==4.25.3
pulsar-client==3.4.0
py-cpuinfo==9.0.0
pyarrow==15.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
PyDispatcher==2.0.6
Pygments==2.17.2
pyOpenSSL==22.1.0
pyparsing==3.1.2
pypdf==4.1.0
PyPDF2==3.0.1
pypdfium2==4.28.0
PyPika==0.48.9
pyproject_hooks==1.0.0
pytesseract==0.3.10
pytest==8.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-poppler==0.4.1
pytz==2024.1
PyYAML==6.0.1
queuelib==1.6.2
rapidfuzz==3.6.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scipy==1.11.1
Scrapy==2.6.3
scrapy-user-agents==0.1.1
Scrapy-UserAgents==0.0.1
service-identity==21.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.36.3
streamlit==1.31.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tesseract==0.1.3
tiktoken==0.6.0
timm==0.9.16
tldextract==3.4.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.0.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.2
transformers==4.39.2
Twisted==22.8.0
typer==0.12.0
typer-cli==0.12.0
typer-slim==0.12.0
types-requests==2.31.0.20240311
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ua-parser==0.16.1
unstructured==0.12.6
unstructured-client==0.18.0
unstructured-inference==0.7.25
unstructured.pytesseract==0.3.12
urllib3==2.2.1
user-agents==2.2.0
uvicorn==0.29.0
uvloop==0.19.0
validators==0.22.0
w3lib==2.0.1
watchfiles==0.21.0
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0
zope.interface==5.5.0
